requirements.txt
# Framework
Django==4.1
djangorestframework==3.14.0

# Lib
psycopg2-binary==2.9.5

# Apps
django-extensions

# Test
factory_boy